requirements.txt
python-telegram-bot==21.6
gspread==6.1.2
google-auth==2.33.0
pytz==2024.1